backend/2/requirements.txt
Django
pytest-django
pytest-cov
factory-boy
pytest-factoryboy
